+++ src/main/java/org/eclipse/keyple/plugin/stub/StubReaderAdapter.java
@@ -17,6 +17,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import org.eclipse.keyple.core.plugin.CardIOException;
 import org.eclipse.keyple.core.plugin.TaskCanceledException;
 import org.eclipse.keyple.core.plugin.spi.reader.ConfigurableReaderSpi;
+import org.eclipse.keyple.core.plugin.spi.reader.PoolReaderSpi;
 import org.eclipse.keyple.core.plugin.spi.reader.observable.ObservableReaderSpi;
 import org.eclipse.keyple.core.plugin.spi.reader.observable.state.insertion.WaitForCardInsertionNonBlockingSpi;
 import org.eclipse.keyple.core.plugin.spi.reader.observable.state.processing.WaitForCardRemovalDuringProcessingBlockingSpi;
@@ -35,6 +36,7 @@ import org.slf4j.LoggerFactory;
 final class StubReaderAdapter
     implements StubReader,
         ConfigurableReaderSpi,
+        PoolReaderSpi,
         ObservableReaderSpi,
         WaitForCardInsertionNonBlockingSpi,
         WaitForCardRemovalDuringProcessingBlockingSpi,
@@ -301,5 +303,15 @@ final class StubReaderAdapter
   @Override
   public void stopWaitForCardRemovalDuringProcessing() {
     continueWaitForCardRemovalTask.set(false);
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @since 2.2.0
+   */
+  @Override
+  public Object getSelectedSmartCard() {
+    return null;
   }
 }