+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added project status badges on `README.md` file.
 ### Fixed
 - CI: code coverage report when releasing.
+### Upgraded
+- "Keyple Plugin API" to version `2.1.0`
+- "Keyple Util Library" to version `2.3.1`
 
 ## [2.1.0] - 2022-06-03
 ### Added